fix/htmlToImage 유틸로 분리
